TODO: need to amend is-website-vulnerable.js to exit with an error code so that the github action will fail when a vulnerability is found... rather than just printing something to the log.
Github action

Uses the existing docker container

The 'scan-url' Github workflow argument is explicity mapped to the
SCAN_URL environment variable already expected&handled by the docker
container.

This enables the Dockerfile to remain agnostic to the fact that it is
running as a Github action, rather than depending on the 'INPUT_'
github action var naming.